pyproject.toml
[project]
name = "simple-local-gallery"
version = "0.1.0"
description = "Simple local gallery"
authors = [
    {name = "Lingyong Sun",email = "lingyong.sun@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.43.2,<2.0.0)",
    "gpxpy (>=1.6.2,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "piexif (>=1.1.3,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "notebook (>=7.3.3,<8.0.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
#
#[tool.poetry]
#dependencies-external = [
#    "python3-tk"
#]
